+++ packages/rocketh/src/environment/utils/extra-chains/index.ts
@@ -1,0 +1,2 @@
+export {ancient8} from './ancient8';
+export {ancient8Sepolia} from './ancient8Sepolia';