+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-deploy
 
+## 4.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.10.8
+
 ## 4.0.9
 
 ### Patch Changes