+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-doc
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.10.8
+
 ## 4.0.7
 
 ### Patch Changes